Move document colors collection to better location.
